Extracted ModelReader<Cell> from Visualiser which is not template now
