+++ visualiser/Visualizer.hpp
@@ -1,12 +1,6 @@
 #pragma once
 
-#include <filesystem>
-#include <iostream>
 #include <string>
-#include <unordered_map>
-#include <format>
-#include <string_view>
-#include <vector>
 #include <vtkActor2D.h>
 #include <vtkCellArray.h>
 #include <vtkCoordinate.h>
@@ -25,71 +19,14 @@
 #include <vtkTextMapper.h>
 #include <vtkTextProperty.h>
 
-#include "Line.h"
 #include "OOpenCAL/base/Element.h" // rgb
-#include "visualiser/SettingParameter.h"
 #include "types.h" // StepIndex
 
-#ifndef __ssize_t_defined
-using ssize_t = intptr_t;
-#endif
+class Line;
 
-
-template <class T>
 class Visualizer
 {
-    std::vector<std::unordered_map<StepIndex, FilePosition>> hashMap;
-
 public:
-    void prepareHashMap(NodeIndex nNodeX, NodeIndex nNodeY)
-    {
-        hashMap.resize(nNodeX * nNodeY);
-    }
-
-    FilePosition getStepStartingPositionInFile(StepIndex step, NodeIndex node)
-    {
-        return hashMap.at(node).at(step);
-    }
-
-    /** @brief Opens the data file for a given simulation step and node.
-     *
-     * The function locates the correct file for the specified node (e.g. "ball3.txt", where 3 is node number),
-     * seeks to the byte position corresponding to the given simulation step in file,
-     * reads the first header line containing local grid dimensions (columns and rows),
-     * and returns an input file stream positioned right after that header.
-     *
-     * @param step         Simulation step number.
-     * @param fileName     Base file name (without node index or extension).
-     * @param node         Node index for which data should be opened.
-     * @param columnAndRow Output: number of local columns and rows read from header line.
-     *
-     * @return std::ifstream Stream ready for reading cell data of this node at given step.
-     * @throws std::runtime_error If the file cannot be opened, seek fails, or header is invalid. */
-    [[nodiscard]] std::ifstream readLocalColumnAndRowForStepFromFileReturningStream(StepIndex step, const std::string& fileName, NodeIndex node, ColumnAndRow& columnAndRow);
-
-    [[nodiscard]] ColumnAndRow readLocalColumnAndRowForStepFromFile(StepIndex step, const std::string& fileName, NodeIndex node);
-
-    template<class Matrix>
-    void readStageStateFromFilesForStep(Matrix& m, SettingParameter* sp, Line *lines);
-
-    /** @brief Loads data into hashMap from text files.
-     *
-     * Expected file format (each line):
-     *     <stepNumber:int> <positionInFile:long>
-     *
-     * Example:
-     *     0 37
-     *     1 32504
-     *     2 64971
-     *
-     * Each line must contain exactly two numbers separated by whitespace.
-     *
-     * @param nNodeX number of nodes along the X axis
-     * @param nNodeY number of nodes along the Y axis
-     * @param filename base filename (e.g., "ball"), for which nodes files are being read */
-    void readStepsOffsetsForAllNodesFromFiles(NodeIndex nNodeX, NodeIndex nNodeY, const std::string &filename);
-
-    /// visualising part:
     template<class Matrix>
     void drawWithVTK(/*const*/ Matrix& p, int nRows, int nCols, StepIndex step, Line *lines, vtkSmartPointer<vtkRenderer> renderer,vtkSmartPointer<vtkActor> gridActor);
     template<class Matrix>
@@ -100,183 +37,14 @@ public:
     vtkNew<vtkActor2D> buildStepText(StepIndex step, int font_size, vtkSmartPointer<vtkNamedColors> colors, vtkSmartPointer<vtkTextProperty> singleLineTextProp, vtkSmartPointer<vtkTextMapper> stepLineTextMapper, vtkSmartPointer<vtkRenderer> renderer);
 
 private:
-    std::pair<std::vector<StepIndex>, std::vector<StepIndex>> giveMeLocalColsAndRowsForAllSteps(StepIndex step, int nNodeX, int nNodeY, const std::string& fileName);
+    template<class Matrix>
+    void buidColor(vtkLookupTable* lut, int nCols, int nRows, Matrix& p);
 };
 
-namespace VisualiserHelpers /// functions which are not templates
-{
-[[nodiscard]] inline std::string giveMeFileName(const std::string& fileName, NodeIndex node)
-{
-    return std::format("{}{}.txt", fileName, node);
-}
-[[nodiscard]] inline std::string giveMeFileNameIndex(const std::string &fileName, NodeIndex node)
-{
-    return std::format("{}{}_index.txt", fileName, node);
-}
-
-ColumnAndRow getColumnAndRowFromLine(const std::string& line);
-
-std::pair<int,int> calculateXYOffset(NodeIndex node, int nNodeX, int nNodeY, const std::vector<int>& allLocalCols, const std::vector<int>& allLocalRows);
-}
 ////////////////////////////////////////////////////////////////////
 
-template <class T>
-ColumnAndRow Visualizer<T>::readLocalColumnAndRowForStepFromFile(StepIndex step, const std::string& fileName, NodeIndex node)
-{
-    ColumnAndRow columnAndRow;
-    std::ifstream file = readLocalColumnAndRowForStepFromFileReturningStream(step, fileName, node, columnAndRow);
-    return columnAndRow;
-}
-template <class T>
-std::ifstream Visualizer<T>::readLocalColumnAndRowForStepFromFileReturningStream(StepIndex step, const std::string& fileName, NodeIndex node, ColumnAndRow &columnAndRow)
-{
-    const auto fileNameTmp = VisualiserHelpers::giveMeFileName(fileName, node);
-
-    std::ifstream file(fileNameTmp);
-    if (! file.is_open())
-    {
-        throw std::runtime_error(std::format("Can't read '{}' in {} function", fileNameTmp, __func__));
-    }
-
-    const auto fPos = getStepStartingPositionInFile(step, node);
-    file.seekg(fPos);
-    if (! file)
-    {
-        throw std::runtime_error(std::format("Seek failed in '{}' at position {}", fileNameTmp, fPos));
-    }
-
-    std::string line;
-    if (! std::getline(file, line))
-    {
-        throw std::runtime_error(std::format("Failed to read line from '{}' at position {}", fileNameTmp, fPos));
-    }
-
-    columnAndRow = VisualiserHelpers::getColumnAndRowFromLine(line);
-    return file;
-}
-
-template<class T>
-template<class Matrix>
-void Visualizer<T>::readStageStateFromFilesForStep(Matrix& m, SettingParameter* sp, Line *lines)
-{
-    const auto [allLocalCols, allLocalRows] = giveMeLocalColsAndRowsForAllSteps(sp->step, sp->nNodeX, sp->nNodeY, sp->outputFileName);
-
-    bool startStepDone = false;
-
-    constexpr std::size_t numbersPerLine = 10'000;
-    const std::size_t lineBufferSize = (log10(UINT_MAX) + 2) * numbersPerLine;
-    std::string line;
-    line.reserve(lineBufferSize); /// for speedup to avoid realocations
-
-    for (NodeIndex node = 0; node < sp->nNodeX * sp->nNodeY; ++node)
-    {
-        const auto [offsetX, offsetY] = VisualiserHelpers::calculateXYOffset(node, sp->nNodeX, sp->nNodeY, allLocalCols, allLocalRows);
-
-        ColumnAndRow columnAndRow;
-        std::ifstream fp = readLocalColumnAndRowForStepFromFileReturningStream(sp->step, sp->outputFileName, node, columnAndRow);
-        if (! fp)
-            throw std::runtime_error("Cannot open file for node " + std::to_string(node));
-
-        /// introducing big buffer for reading (for speed up)
-        static thread_local char fileBuffer[1 << 16];
-        fp.rdbuf()->pubsetbuf(fileBuffer, sizeof(fileBuffer));
-
-        lines[node * 2]     = Line(offsetX, offsetY, offsetX + columnAndRow.column, offsetY);
-        lines[node * 2 + 1] = Line(offsetX, offsetY, offsetX, offsetY + columnAndRow.row);
-
-        for (int row = 0; row < columnAndRow.row; ++row)
-        {
-            if (! std::getline(fp, line))
-            {
-                const auto fileNameTmp = VisualiserHelpers::giveMeFileName(sp->outputFileName, node);
-                throw std::runtime_error("Error when reading entire line of file " + fileNameTmp);
-            }
-
-            std::replace(begin(line), end(line), ' ', '\0'); /// this is faster than using std::ranges::replace
-
-            /// Moving through tokens (token are characters ended with '0')
-            char* currentTokenPtr = line.data();
-            for (int col = 0; col < columnAndRow.column && *currentTokenPtr; ++col)
-            {
-                if (! startStepDone) [[unlikely]]
-                {
-                    m[row + offsetY][col + offsetX].T::startStep(sp->step);
-                    startStepDone = true;
-                }
-
-                m[row + offsetY][col + offsetX].T::composeElement(currentTokenPtr);
-
-                /// go to another token
-                while (*currentTokenPtr)
-                    ++currentTokenPtr;
-                ++currentTokenPtr; /// skip '\0'
-            }
-        }
-    }
-}
-template<class T>
-std::pair<std::vector<StepIndex>, std::vector<StepIndex>> Visualizer<T>::giveMeLocalColsAndRowsForAllSteps(StepIndex step, int nNodeX, int nNodeY, const std::string& fileName)
-{
-    const auto nodesCount = nNodeX * nNodeY;
-    std::vector<StepIndex> allLocalCols{nodesCount}, allLocalRows{nodesCount};
-    allLocalCols.resize(nodesCount);
-    allLocalRows.resize(nodesCount);
-
-    for (NodeIndex node = 0; node < nodesCount; node++)
-    {
-        auto [nLocalCols, nLocalRows] = readLocalColumnAndRowForStepFromFile(step, fileName, node);
-
-        allLocalCols[node] = nLocalCols;
-        allLocalRows[node] = nLocalRows;
-    }
-    return {allLocalCols, allLocalRows};
-}
-
-template <class T>
-void Visualizer<T>::readStepsOffsetsForAllNodesFromFiles(NodeIndex nNodeX, NodeIndex nNodeY, const std::string& filename)
-{
-    const auto totalNodes = nNodeX * nNodeY;
-    for (NodeIndex node = 0; node < totalNodes; ++node)
-    {
-        const auto fileNameIndex = VisualiserHelpers::giveMeFileNameIndex(filename, node);
-        std::cout << "Reading file: " << fileNameIndex << '\n';
-
-        if (! std::filesystem::exists(fileNameIndex))
-        {
-            throw std::runtime_error("File not found: " + fileNameIndex);
-        }
-
-        std::ifstream file(fileNameIndex);
-        if (! file)
-        {
-            throw std::runtime_error("Cannot open file: " + fileNameIndex);
-        }
-
-        while (true)
-        {
-            StepIndex stepNumber;
-            FilePosition positionInFile;
-
-            if (! (file >> stepNumber >> positionInFile))
-            { /// enters here if reading error
-                if (file.eof())
-                    break;
-                else
-                    throw std::runtime_error("Invalid line format in file: " + fileNameIndex);
-            }
-
-            const auto [it, inserted] = hashMap[node].emplace(stepNumber, positionInFile);
-            if (! inserted)
-            {
-                std::cerr << std::format("Duplicate stepNumber {} found in file '{}' (node {})", stepNumber, fileNameIndex, node) << std::endl;
-            }
-        }
-    }
-}
-
-template <class T>
 template <class Matrix>
-void Visualizer<T>::drawWithVTK(/*const*/ Matrix& p, int nRows, int nCols, StepIndex step, Line *lines, vtkSmartPointer<vtkRenderer> renderer, vtkSmartPointer<vtkActor> gridActor)
+void Visualizer::drawWithVTK(/*const*/ Matrix& p, int nRows, int nCols, StepIndex step, Line *lines, vtkSmartPointer<vtkRenderer> renderer, vtkSmartPointer<vtkActor> gridActor)
 {
     const auto numberOfPoints = nRows * nCols;
     vtkNew<vtkDoubleArray> pointValues;
@@ -315,9 +83,8 @@ void Visualizer<T>::drawWithVTK(/*const*/ Matrix& p, int nRows, int nCols, StepI
     renderer->AddActor(gridActor);
 }
 
-template <class T>
 template <class Matrix>
-void Visualizer<T>::refreshWindowsVTK(/*const*/ Matrix& p, int nRows, int nCols, StepIndex step, Line *lines, int dimLines, vtkSmartPointer<vtkActor> gridActor)
+void Visualizer::refreshWindowsVTK(/*const*/ Matrix& p, int nRows, int nCols, StepIndex step, Line *lines, int dimLines, vtkSmartPointer<vtkActor> gridActor)
 {
     if (vtkLookupTable* lut = dynamic_cast<vtkLookupTable*>(gridActor->GetMapper()->GetLookupTable()))
     {
@@ -329,110 +96,8 @@ void Visualizer<T>::refreshWindowsVTK(/*const*/ Matrix& p, int nRows, int nCols,
         throw std::runtime_error("Invalid dynamic cast!");
 }
 
-template <class T>
-void Visualizer<T>::buildLoadBalanceLine(Line *lines, int dimLines,int nCols, int nRows, vtkSmartPointer<vtkPoints> pts, vtkSmartPointer<vtkCellArray> cellLines, vtkSmartPointer<vtkPolyData> grid, vtkSmartPointer<vtkNamedColors> colors, vtkSmartPointer<vtkRenderer> renderer,vtkSmartPointer<vtkActor2D> actorBuildLine)
-{
-    for (int i = 0; i < dimLines; i++)
-    {
-        std::cout << "Line (" << lines[i].x1 << ", " << lines[i].y1 << ") -> (" <<lines[i].x2 << ", " <<lines[i].y2 << ")" << std::endl;
-        pts->InsertNextPoint(lines[i].x1 * 1, nCols-1-lines[i].y1 * 1, 0.0);
-        pts->InsertNextPoint(lines[i].x2 * 1, nCols-1-lines[i].y2 * 1, 0.0);
-        cellLines->InsertNextCell(2);
-        cellLines->InsertCellPoint(i*2);
-        cellLines->InsertCellPoint(i*2+1);
-    }
-
-    grid->SetPoints(pts);
-    grid->SetLines(cellLines);
-    // Set up the coordinate system.
-    vtkNew<vtkCoordinate> normCoords;
-    normCoords->SetCoordinateSystemToWorld();
-
-    vtkNew<vtkPolyDataMapper2D> mapper;
-    mapper->SetInputData(grid);
-    mapper->Update();
-    mapper->SetTransformCoordinate(normCoords);
-
-    actorBuildLine->SetMapper(mapper);
-    actorBuildLine->GetMapper()->Update();
-    actorBuildLine->GetProperty()->SetColor(colors->GetColor3d("Red").GetData());
-   // actorBuildLine->GetProperty()->SetLineWidth(1.5);
-    renderer->AddActor2D(actorBuildLine);
-}
-
-template <class T>
-void Visualizer<T>::refreshBuildLoadBalanceLine(Line *lines, int dimLines,int nCols,int nRows, vtkActor2D* lineActor,vtkSmartPointer<vtkNamedColors> colors)
-{
-    vtkPolyDataMapper2D* mapper = (vtkPolyDataMapper2D*) lineActor->GetMapper();
-    mapper->Update();
-    vtkNew<vtkPolyData> grid;
-    vtkNew<vtkPoints> pts;
-    vtkNew<vtkCellArray> cellLines;
-
-    for (int i = 0; i < dimLines; i++)
-    {
-        pts->InsertNextPoint((lines[i].x1 * 1), ( nCols-1-lines[i].y1 * 1), 0.0);
-        pts->InsertNextPoint((lines[i].x2 * 1), ( nCols-1-lines[i].y2 * 1), 0.0);
-        cellLines->InsertNextCell(2);
-        cellLines->InsertCellPoint(i*2);
-        cellLines->InsertCellPoint(i*2+1);
-    }
-
-    grid->SetPoints(pts);
-    grid->SetLines(cellLines);
-    // Set up the coordinate system.
-    vtkNew<vtkCoordinate> normCoords;
-    normCoords->SetCoordinateSystemToWorld();
-
-    mapper->SetInputData(grid);
-    mapper->Update();
-    mapper->SetTransformCoordinate(normCoords);
-    
-    lineActor->SetMapper(mapper);
-    lineActor->GetMapper()->Update();
-}
-
-template <class T>
-vtkTextProperty* Visualizer<T>::buildStepLine(StepIndex step, vtkSmartPointer<vtkTextMapper> singleLineTextB, vtkSmartPointer<vtkTextProperty> singleLineTextProp, vtkSmartPointer<vtkNamedColors> colors, std::string color)
-{
-    std::string stepText = "Step " + std::to_string(step);
-    singleLineTextB->SetInput(stepText.c_str());
-    singleLineTextB->Update();
-    vtkTextProperty* tprop = singleLineTextB->GetTextProperty();
-    tprop->ShallowCopy(singleLineTextProp);
-    tprop->SetVerticalJustificationToBottom();
-    tprop->SetColor(colors->GetColor3d(color).GetData());
-    return tprop;
-}
-
-
-template <class T>
-vtkNew<vtkActor2D> Visualizer<T>::buildStepText(StepIndex step, int font_size, vtkSmartPointer<vtkNamedColors> colors, vtkSmartPointer<vtkTextProperty> singleLineTextProp, vtkSmartPointer<vtkTextMapper> stepLineTextMapper, vtkSmartPointer<vtkRenderer> renderer)
-{
-    singleLineTextProp->SetFontSize(font_size);
-    singleLineTextProp->SetFontFamilyToArial();
-    singleLineTextProp->BoldOn();
-    singleLineTextProp->ItalicOff();
-    singleLineTextProp->ShadowOff();
-
-    const std::string stringWithStep = "Step " + std::to_string(step);
-    stepLineTextMapper->SetInput(stringWithStep.c_str());
-    auto tprop = stepLineTextMapper->GetTextProperty();
-    tprop->ShallowCopy(singleLineTextProp);
-
-    tprop->SetVerticalJustificationToBottom();
-    tprop->SetColor(colors->GetColor3d("Red").GetData());
-
-    vtkNew<vtkActor2D> stepLineTextActor;
-    stepLineTextActor->SetMapper(stepLineTextMapper);
-    stepLineTextActor->GetPositionCoordinate()->SetCoordinateSystemToNormalizedDisplay();
-    stepLineTextActor->GetPositionCoordinate()->SetValue(0.05, 0.85);
-    renderer->AddActor2D(stepLineTextActor);
-    return stepLineTextActor;
-}
-
 template <class Matrix>
-void buidColor(vtkLookupTable* lut, int nCols, int nRows, Matrix& p)
+void Visualizer::buidColor(vtkLookupTable* lut, int nCols, int nRows, Matrix& p)
 {
     for (int r = 0; r < nRows; ++r)
     {
